# png decode perf improvement