before:
`pixie decode ....................... 4.836 ms      4.904 ms    ±0.043  x1000`

after:
`pixie decode ....................... 4.312 ms      4.360 ms    ±0.031  x1000`